backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
pydantic>=2.0.0
# Optional fallback for transcript extraction
youtube-transcript-api>=0.6.0 